+++ src/App.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 import './App.css';
 import { BrowserRouter, Navigate, Route, Routes } from 'react-router-dom';
 import { Error404 } from './pages/Error404';
-import { AboutUs } from './pages/AboutUs';
-import { PostsPage } from './pages/PostsPage';
+import { AboutUs } from './pages/AboutUs/AboutUs';
+import { PostsPage } from './pages/PostsPage/PostsPage';
 
 function App() {
   return (
