+++ src/Pages/AboutUs/AboutUs.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
-import { MyHeader } from '../components/MyHeader/MyHeader'
+import { MyHeader } from '../../components/MyHeader/MyHeader'
 
 export function AboutUs() {
   return (
   <div>
-    <MyHeader></MyHeader>
+    <MyHeader active={'about'}></MyHeader>
     <h1>Hello! Sorry, there is nothing special here yet, but it may appear later</h1>;
   </div>)
 }
