+++ src/Pages/PostsPage/PostsPage.module.css
@@ -1,0 +1,13 @@
+.posts_container{
+    display: grid;
+    grid-template-rows: 60vmin;
+    grid-template-columns: repeat(auto-fill, 40vmin);
+    grid-gap: 5vmin;
+    justify-content: center;
+    margin: 3vmin 0;
+}
+
+h1, h2, h3{
+    font-weight: 100;
+    margin: 0;
+}
