+++ src/components/MyInput/MyInput.module.css
@@ -1,0 +1,8 @@
+.input{
+    padding-left: 1vmin;
+    border: none;
+    border-radius: 1vmin;
+    font-family: system-ui;
+    font-weight: 100;
+    font-size: 3.5vmin;
+}
