+++ src/index.css
@@ -1,1 +1,6 @@
-
+body{
+    font-family: system-ui;
+    font-style: normal;
+    font-weight: 100;
+    margin: 0;
+}
